first pass at carving out vehicle-import feature
